FIX: CAR-68 메인 페이지 공고 오류 해결 (2025-04-28)
