+++ src/main/java/com/highfive/meetu/global/config/SecurityConfig.java
@@ -46,10 +46,11 @@ public class SecurityConfig {
             .authorizeHttpRequests(authorize -> authorize
                 .requestMatchers("/api/admin/dashboard/**").hasRole("ADMIN")  // 관리자 대시보드 보호
                 .requestMatchers(
-                    "/api/admin/auth/login", // 관리자 로그인
-                    "/api/business/auth/login", // 기업회원 로그인
-                    "/api/personal/auth/login", // 개인회원 로그인
-                    "/api/main/**",             // 메인
+                    "/api/admin/auth/login",        // 관리자 로그인
+                    "/api/business/auth/login",     // 기업회원 로그인
+                    "/api/personal/auth/login",     // 개인회원 로그인
+                    "/api/main/**",                 // 메인
+                    "/api/personal/job/**",       // 채용 정보
                     "/swagger-ui/**",
                     "/v3/api-docs/**",
                     "/error"
